+++ bakery_model/Cargo.toml
@@ -16,6 +16,7 @@ testcontainers-modules = { version = "0.8.0", features = [
 tokio = "1.38.1"
 tokio-postgres = "0.7.10"
 sqlformat = "0.2.3"
+sqlx = { version = "0.8.2", default-features = false, features = ["json", "postgres", "runtime-tokio"] }
 
 [[example]]
 name = "0-intro"
