+++ bakery_model/src/lib.rs
@@ -44,32 +44,6 @@ pub async fn connect_postgres() -> Result<()> {
     let connection_string = std::env::var("DATABASE_URL")
         .unwrap_or_else(|_| "postgres://postgres@localhost:5432/postgres".to_string());
 
-    let timeout = Duration::from_secs(3); // Max time to wait
-    let start_time = Instant::now();
-    let mut last_error: Result<()> = Ok(());
-
-    while Instant::now().duration_since(start_time) < timeout {
-        match tokio_postgres::connect(&connection_string, NoTls).await {
-            Ok((client, connection)) => {
-                tokio::spawn(async move {
-                    if let Err(e) = connection.await {
-                        eprintln!("connection error: {}", e);
-                    }
-                });
-
-                set_postgres(Postgres::new(Arc::new(Box::new(client))))?;
-
-                println!("Successfully connected to the database.");
-                return Ok(());
-            }
-            Err(e) => {
-                println!("Error connecting to database: {}, retrying...", &e);
-                last_error = Err(anyhow::Error::new(e));
-                // sleep(Duration::from_secs(2)).await; // Wait before retrying
-                thread::sleep(Duration::from_millis(100));
-            }
-        }
-    }
-
-    last_error
+    let postgres = Postgres::new(&connection_string).await;
+    set_postgres(postgres)
 }
