+++ bakery_model/tests/bakery_integration_test.rs
@@ -62,11 +62,11 @@ async fn create_bootstrap_db() -> Result<()> {
 
     // Get the postgres client for batch execution
     let vantage_client = bakery_model::postgres();
-    let client = vantage_client.client();
+    let client = vantage_client.pool;
 
     // Read the schema from the file and execute it
     let schema = tokio::fs::read_to_string("schema-pg.sql").await?;
-    client.batch_execute(&schema).await?;
+    sqlx::raw_sql(&schema).execute(&*client).await?;
 
     Ok(())
 }
