+++ vantage/src/datasource/mod.rs
@@ -1,2 +1,3 @@
+pub mod associated_query;
 pub mod postgres;
 pub mod sqlx;
