+++ vantage/src/prelude.rs
@@ -1,6 +1,8 @@
 pub use crate::dataset::ReadableDataSet;
 pub use crate::dataset::WritableDataSet;
-pub use crate::datasource::postgres::*;
+// pub use crate::datasource::postgres::*;
+pub use crate::datasource::associated_query::AssociatedQuery;
+pub use crate::datasource::sqlx::postgres::*;
 pub use crate::expr;
 pub use crate::expr_arc;
 pub use crate::mocks::MockDataSource;
